migrating tests to pytest, adding flaky to checkrunner
