+++ .travis.yml
@@ -7,7 +7,7 @@ install:
 - pip install coverage
 - pip install coveralls
 script:
-- coverage run test.py
+- pytest tests
 - coverage report -m
 after_success:
 - coveralls
