+++ requirements.txt
@@ -7,4 +7,6 @@ MarkupSafe==1.0
 google-api-python-client==1.7.4
 geocoder==1.38.1
 fuzzywuzzy==0.17.0
+pytest
+flaky
 -e git+git://github.com/4dn-dcic/chalice.git#egg=chalice
