+++ tests/test_fs_connection.py
@@ -1,0 +1,69 @@
+import chalice
+import datetime
+import json
+import boto3
+import app
+from chalicelib import (
+    app_utils,
+    check_utils,
+    utils,
+    run_result,
+    fs_connection,
+    s3_connection
+)
+from dcicutils import s3_utils, ff_utils
+import pytest
+
+
+class TestFSConnection():
+    environ_info = {
+        'fourfront': 'test1',
+        'es': 'test2',
+        'bucket': None,
+        'ff_env': 'test3'
+    }
+    connection = fs_connection.FSConnection('test', environ_info, test=True)
+
+    def test_connection_fields(self):
+        assert (self.connection.fs_env == 'test')
+        assert (self.connection.s3_connection.status_code == 404)
+        assert (self.connection.ff_server == 'test1')
+        assert (self.connection.ff_es == 'test2')
+        assert (self.connection.ff_env == 'test3')
+        assert (self.connection.ff_s3 is None)
+        assert (self.connection.ff_keys is None)
+
+    def test_run_check_with_bad_connection(self):
+        check_res = check_utils.run_check_or_action(self.connection, 'wrangler_checks/item_counts_by_type', {})
+        # run_check_or_action returns a dict with results
+        assert (check_res.get('status') == 'ERROR')
+        assert (check_res.get('name') == 'item_counts_by_type')
+
+    def test_check_result_basics(self):
+        test_check = utils.init_check_res(self.connection, 'test_check')
+        test_check.summary = 'Unittest check'
+        test_check.ff_link = 'not_a_real_http_link'
+        assert (test_check.s3_connection.status_code == 404)
+        assert (test_check.get_latest_result() is None)
+        assert (test_check.get_primary_result() is None)
+        with pytest.raises(Exception) as exec_info:
+            test_check.get_closest_result(1)
+        assert ('Could not find any results' in str(exec_info.value))
+        formatted_res = test_check.format_result(datetime.datetime.utcnow())
+        assert (formatted_res.get('status') == 'IGNORE')
+        assert (formatted_res.get('summary') == 'Unittest check')
+        assert (formatted_res.get('description') == 'Unittest check')
+        assert (formatted_res.get('type') == 'check')
+        # set a bad status on purpose
+        test_check.status = "BAD_STATUS"
+        check_res = test_check.store_result()
+        assert (check_res.get('name') == formatted_res.get('name'))
+        assert (check_res.get('description') == "Malformed status; look at Foursight check definition.")
+        assert (check_res.get('brief_output') == formatted_res.get('brief_output') == None)
+        assert (check_res.get('ff_link') == 'not_a_real_http_link')
+
+    def test_bad_ff_connection_in_fs_connection(self):
+        # do not set test=True, should raise because it's not a real FF
+        with pytest.raises(Exception) as exec_info:
+            bad_connection = fs_connection.FSConnection('test', self.environ_info)
+        assert ('Could not initiate connection to Fourfront' in str(exec_info.value))
